Minor changes to get rid of compiler warnings on 64-bit linux

+++ emu/debug_xsb.c
@@ -516,7 +516,7 @@ static void print_term_of_subgoal(CTXTdeclc FILE *fp, int *i)
   /*    fprintf(fp, "%s", string_val(term));*/
     break;
   case XSB_INT:
-    fprintf(fp, "%d", int_val(term));
+    fprintf(fp, "%d", (int) int_val(term));
     break;
   case XSB_FLOAT:
     fprintf(fp, "%.5g", float_val(term));
